Namespaced life events
