+++ app/helpers/marriage_helper.rb
@@ -1,0 +1,2 @@
+module MarriageHelper
+end
