+++ app/views/life_event/marriages/edit.html.erb
@@ -1,0 +1,8 @@
+<% provide(:title,
+  "Editing the marriage between #{@marriage.person_1} and #{@marriage.person_2}") %>
+<h1>Editing marriage</h1>
+
+<%= render 'form' %>
+
+<%= link_to 'Show', @person %> |
+<%= link_to 'Back', people_path %>
