+++ app/views/life_event/marriages/new.html.erb
@@ -1,0 +1,5 @@
+<h1>New marriage</h1>
+
+<%= render 'form' %>
+
+<%= link_to 'Back', people_path %>
