+++ spec/controllers/marriage_controller_spec.rb
@@ -1,0 +1,5 @@
+require 'rails_helper'
+
+RSpec.describe MarriageController, type: :controller do
+
+end
