--- spec/factories/birth.rb
@@ -1,5 +1,0 @@
-FactoryGirl.define do
-  factory :birth do
-    date { Faker::Date.between(99.years.ago, Date.yesterday) }
-  end
-end
