+++ spec/factories/life_event/birth.rb
@@ -1,0 +1,5 @@
+FactoryGirl.define do
+  factory :birth, class: "LifeEvent::Birth" do
+    date { Date.yesterday }
+  end
+end
