+++ spec/factories/life_event/death.rb
@@ -1,0 +1,5 @@
+FactoryGirl.define do
+  factory :death, class: "LifeEvent::Death" do
+    date { Date.yesterday }
+  end
+end
