--- spec/factories/marriages.rb
@@ -1,6 +1,0 @@
-FactoryGirl.define do
-  factory :marriage do
-    person_1 { build(:person) }
-    person_2 { build(:person) }
-  end
-end
