+++ spec/models/life_event/death_spec.rb
@@ -1,9 +1,7 @@
 require 'spec_helper'
 
-describe Death do
-  before { @death = Death.create() }
-
-  subject { @death }
+describe LifeEvent::Death, type: :model do
+  subject(:death) { build(:death) }
 
   it { is_expected.to respond_to(:date)}
   it { is_expected.to respond_to(:place)}
@@ -12,19 +10,17 @@ describe Death do
   it { is_expected.to respond_to(:dead)}
 
   context "when date is in the future" do
-    before { @death.date = Date.tomorrow }
+    before { death.date = Date.tomorrow }
 
     it { is_expected.to be_invalid }
   end
 
   describe "#dead" do
     context "when date is in the past" do
-      before do
-        @death.date = Date.new(1900,01,01)
-      end
+      before { death.date = Date.new(1900,01,01) }
 
       specify "is true" do
-        expect(@death.dead).to be_truthy
+        expect(death.dead).to be_truthy
       end
     end
   end
