+++ spec/models/life_event/marriage_spec.rb
@@ -1,6 +1,6 @@
 require 'spec_helper'
 
-RSpec.describe Marriage, type: :model do
+describe LifeEvent::Marriage, type: :model do
   subject(:marriage) { build(:marriage) }
 
   it { is_expected.to respond_to(:person_1) }
